+++ tools/AssetLib/src/ModelAsset.h
@@ -60,7 +60,7 @@ namespace assetLib
     {
         static constexpr u32 TRIANGLES_PER_MESHLET = 256;
         static constexpr u32 VERTICES_PER_MESHLET = 255;
-        using IndexType = u16;
+        using IndexType = u8;
         enum class MaterialType : u32
         {
             Albedo = 0, MaxVal
